Moved out of the __init__.py the functions: handle_achievements() and check_and_award_badges()
